Add setup.py and rename "libs" to "distaf"
In order to make it possible to package DiSTAF and include it in
repositories of OS distributions, some modifications to the directory
structure are needed:
 - rename the "libs" directory to "distaf"
 - move "main.py" to the new "distaf" directory
 - symlink "distaf/main.py" to "main.py" to keep current workflow
 - correct the "import" statements that used the "libs" packages

A "__version__" has also been added to "main.py". The new "setup.py" can
be used to create a tarball for distributing released versions.
