+++ MANIFEST.in
@@ -1,0 +1,4 @@
+include README.md
+include CONTRIBUTING
+include LICENSE
+include config.sh
